Changes: - Fixed typo on Create new Catalog object with new Type - Added sample commit controller method for the Catalog Service - Added Token parsing to Catalog service's TCC middleware - Added Proposal Timestamp Logic for the Catalog Service - Added Commit issuing logic to Coordinator Service, sent to services that participated in the functionality

+++ src/Services/Coordinator/Coordinator.API/Services/IFunctionalityService.cs
@@ -3,10 +3,10 @@
 namespace Microsoft.eShopOnContainers.Services.Coordinator.API.Services;
 public interface IFunctionalityService {
     ConcurrentDictionary<string, List<(string, long)>> Proposals { get; }
-    ConcurrentDictionary<string, int> Tokens { get; }
+    ConcurrentDictionary<string, double> Tokens { get; }
 
     public void AddNewProposalGivenService(string funcID, string service, long proposalTicks);
-    public void IncreaseTokens(string funcID, int tokens);
+    public void IncreaseTokens(string funcID, double tokens);
     public Boolean HasCollectedAllTokens(string funcID);
 
 }
